refactor: simplify user action button
