--- apps/web/components/buttons/FloatingUserButton/signed-in.tsx
@@ -1,50 +1,0 @@
-import { UserIcon } from '@/ui/icons'
-import { useUser } from '@/utils/auth0'
-import { DropdownMenu, Entries } from '@/ui/menu'
-import { useAppDispatch, useAppSelector } from '@/store/redux'
-import { hideContextMenu, showDebugModal, showUserProfile } from '@/store/modal'
-import { MenuItem } from 'types'
-
-export default function SignedInCircle() {
-  const { user } = useUser()
-  const dispatch = useAppDispatch()
-  const dev = process.env['NODE_ENV'] === 'development'
-  const email = useAppSelector((state) => state.user.email)
-
-  const items: MenuItem[] = [
-    { text: 'Your profile', callback: () => dispatch(showUserProfile()) },
-    {
-      text: 'Debug',
-      callback: () => dispatch(showDebugModal()),
-    },
-    { text: 'Settings', callback: () => alert('open settings') },
-    {
-      text: 'Sign out',
-      href: '/api/auth/logout',
-    },
-  ]
-
-  const UserCircle = () => (
-    <div
-      className="w-10 h-10 rounded-full modtree-gradient flex centered"
-      onClick={() => dispatch(hideContextMenu())}
-    >
-      <UserIcon className="text-gray-50 h-5 w-5" />
-    </div>
-  )
-
-  return (
-    <DropdownMenu TriggerButton={UserCircle}>
-      <Entries items={items.filter((item) => dev || item.text !== 'Debug')}>
-        <>
-          <div>Signed in as{dev && ' (auth0)'}</div>
-          <b>{user.nickname}</b>
-        </>
-        <>
-          <div>Email{dev && ' (postgres)'}</div>
-          <b>{email}</b>
-        </>
-      </Entries>
-    </DropdownMenu>
-  )
-}
